Add address & education

+++ models/education.go
@@ -6,7 +6,7 @@ type Education struct {
 	PersonID    string   `gorm:"not null"`
 	Institution string   `gorm:"not null"`
 	Subject     string   `gorm:"not null"`
-	Grade       *float32 `gorm:"not null"`
+	Grade       *float64 `gorm:"not null"`
 	YearBegin   *int     ``
 	YearEnd     *int     ``
 	CreatedBy   string   `gorm:"not null;default:admin"`
@@ -19,17 +19,3 @@ type Education struct {
 func (Education) TableName() string {
 	return "cor_person_educations"
 }
-
-// // ConvertFromRequest ...
-// func (Person) ConvertFromRequest(r *requests.PersonUpdateRequest) *Person {
-// 	return &Person{
-// 		FirstName:         r.FirstName,
-// 		LastName:          r.LastName,
-// 		Email:             r.Email,
-// 		BirthPlace:        r.BirthPlace,
-// 		BirthDate:         datatypes.Date(r.BirthDate),
-// 		Mobile:            r.Mobile,
-// 		Gender:            r.Gender,
-// 		IsFromRecruitment: r.IsFromRecruitment,
-// 	}
-// }
